# server/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World of Stories API",
    "version": "1.0.0",
    "description": "Documentación limpia y modularizada"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/places/city/{cityId}": {
      "get": {
        "operationId": "getCityPlaces",
        "tags": [
          "Places"
        ],
        "summary": "Obtener todos los lugares de una ciudad (getCityPlaces)",
        "description": "Retorna los lugares que pertenecen a una ciudad específica. Soporta filtrado por el query param ?published=true.",
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "required": true,
            "description": "ID de la ciudad a la que pertenecen los lugares",
            "schema": {
              "type": "string",
              "example": "665ed7c4c3f4e567890defab"
            }
          },
          {
            "name": "published",
            "in": "query",
            "required": false,
            "description": "Filtrar lugares por estado de publicación (?published=true)",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lugares de la ciudad obtenida correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 3
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "places": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Place"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró la ciudad especificada."
          }
        }
      },
      "post": {
        "operationId": "createPlace",
        "tags": [
          "Places"
        ],
        "summary": "Crear y guardar un nuevo lugar en una ciudad (savePlace)",
        "description": "Inserta un nuevo lugar dentro del array de lugares de la ciudad indicada.",
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Palacio de Elendel"
                  },
                  "userName": {
                    "type": "string",
                    "example": "Palacio de Elendel"
                  },
                  "description": {
                    "type": "Object",
                    "$ref": "#/components/schemas/Readable"
                  },
                  "entry": {
                    "type": "Object",
                    "$ref": "#/components/schemas/Readable"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lugar creado y guardado de forma exitosa.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Lugar guardado correctamente."
                    },
                    "data": {
                      "$ref": "#/components/schemas/Place"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "La ciudad especificada no existe."
          }
        }
      }
    },
    "/places/{placeId}": {
      "get": {
        "operationId": "getPlace",
        "tags": [
          "Places"
        ],
        "summary": "Obtener detalle completo de un solo lugar (getOnePlace)",
        "description": "Busca en toda la base de datos y retorna un único lugar por su ID. Devuelve todos los campos desglosados (events como array de IDs, descriptions y entries formateados).",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "description": "ID único del lugar a buscar",
            "schema": {
              "type": "string",
              "example": "665ed5a1c3f4e567890bcdef"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del lugar recuperado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "place": {
                          "$ref": "#/components/schemas/Place"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró el lugar especificado."
          }
        }
      },
      "patch": {
        "operationId": "updatePlace",
        "tags": [
          "Places"
        ],
        "summary": "Actualizar un lugar (updatePlace)",
        "description": "Modifica de manera parcial el estado de publicación y setea automáticamente el campo publishDate.",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "place"
                ],
                "properties": {
                  "place": {
                    "$ref": "#/components/schemas/Place"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado de publicación alterado con éxito."
          }
        }
      },
      "delete": {
        "operationId": "deletePlace",
        "tags": [
          "Places"
        ],
        "summary": "Eliminar un lugar por completo (deletePlace)",
        "description": "Remueve de forma definitiva el subdocumento del lugar sin requerir el ID de la ciudad.",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lugar eliminado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Lugar eliminado correctamente."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lugar no encontrado."
          }
        }
      }
    },
    "/places/{placeId}/publish": {
      "patch": {
        "operationId": "publishPlace",
        "tags": [
          "Places"
        ],
        "summary": "Publicar o despublicar un lugar (publishPlace)",
        "description": "Modifica de manera parcial el estado de publicación y setea automáticamente el campo publishDate.",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "published"
                ],
                "properties": {
                  "published": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado de publicación alterado con éxito."
          }
        }
      }
    },
    "/places/{placeId}/description": {
      "put": {
        "operationId": "addPlaceDescription",
        "tags": [
          "Places"
        ],
        "summary": "Añadir un relato de descripción a un lugar (addPlaceDescription)",
        "description": "Inserta una nueva descripción en el array del lugar.",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tale",
                  "author"
                ],
                "properties": {
                  "tale": {
                    "type": "Object",
                    "$ref": "#/components/schemas/Readable"
                  },
                  "author": {
                    "type": "string",
                    "example": "Hoid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Descripción guardada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "description": {
                          "$ref": "#/components/schemas/Tale"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updatePlaceDescription",
        "tags": [
          "Places"
        ],
        "summary": "Actualizar un relato de descripción específico (updatePlaceDescription)",
        "description": "Actualiza de forma atómica usando arrayFilters el relato interno del lugar.",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "description"
                ],
                "properties": {
                  "description": {
                    "type": "object",
                    "required": [
                      "id",
                      "tale",
                      "published"
                    ],
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "665ed1b2c3f4e567890abcde"
                      },
                      "tale": {
                        "type": "Object",
                        "$ref": "#/components/schemas/Readable"
                      },
                      "published": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Descripción actualizada con éxito."
          }
        }
      }
    },
    "/places/{placeId}/description/{taleId}": {
      "delete": {
        "operationId": "removePlaceDescription",
        "tags": [
          "Places"
        ],
        "summary": "Eliminar una descripción del lugar (removePlaceDescription)",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Descripción removida del array satisfactoriamente."
          }
        }
      }
    },
...
    "/places/{placeId}/entry/{taleId}": {
      "delete": {
        "operationId": "removePlaceEntry",
        "tags": [
          "Places"
        ],
        "summary": "Eliminar un cuento de entrada del lugar (removePlaceEntry)",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entrada del lugar eliminada con éxito."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Readable": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "El aire soplaba frío sobre las llanuras quebradas..."
          },
          "animation": {
            "type": "string",
            "example": "fade-in-slow"
          }
        }
      },
...
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665ed5a1c3f4e567890bcdef"
          },
          "name": {
            "type": "string",
            "example": "Kholinar"
          },
          "published": {
            "type": "boolean",
            "example": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-15T00:00:00.000Z"
          },
          "events": {
            "type": "array",
            "description": "IDs de los NPCs vinculados a este lugar",
            "items": {
              "type": "string",
              "example": "665ed6b3c3f4e567890cdefa"
            }
          },
          "description": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tale"
            }
          },
          "entry": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tale"
            }
          }
        }
      },
...
    }
  }
}